fix: Checking the state of a database with a binary field.
