+++ tests/fixtures/TableTestStateTest/db_changes/test_models/null_to_binary_resource_changes.json
